+++ domain/src/main/java/com/weatherapp/domain/usecases/GetWeatherByLocationApiUseCase.kt
@@ -8,7 +8,7 @@ import com.weatherapp.data.repository.WeatherService
 
 class GetWeatherByLocationApiUseCase(private val service: WeatherService): UseCase<WeatherEntity, GetWeatherByLocationApiUseCase.Params>() {
 
-    data class Params(val lat: Long, val lon: Long)
+    data class Params(val lat: Double, val lon: Double)
 
     override suspend fun run(params: Params): Either<Failure, WeatherEntity> {
         return service.getWeatherByCoordinate(params.lat, params.lon)
